+++ lib/ca_testing/drivers/v4/options.rb
@@ -1,16 +1,9 @@
 # frozen_string_literal: true
 
-require "selenium/webdriver/common/options"
-require "selenium/webdriver/chrome/options"
-require "selenium/webdriver/firefox/options"
-require "selenium/webdriver/safari/options"
-
-# This is needed due to the paths being inconsistent during the v4 alphas/betas
-begin
-  require "selenium/webdriver/edge/options"
-rescue LoadError
-  require "selenium/webdriver/edge_chrome/options"
-end
+require "selenium/webdriver/chrome"
+require "selenium/webdriver/firefox"
+require "selenium/webdriver/safari"
+require "selenium/webdriver/edge"
 
 module CaTesting
   module Drivers
@@ -30,19 +23,10 @@ module CaTesting
             case browser
             when :chrome;             then ::Selenium::WebDriver::Chrome::Options.new
             when :firefox;            then ::Selenium::WebDriver::Firefox::Options.new(log_level: "trace")
-            when :edge;               then edge_options
+            when :edge;               then ::Selenium::WebDriver::Edge::Options.new
             when :safari;             then ::Selenium::WebDriver::Safari::Options.new(automatic_inspection: true)
             when :internet_explorer;  then internet_explorer_options
             else {}
-            end
-          end
-
-          # This is only needed before version 4.0 is stable
-          def edge_options
-            if defined? ::Selenium::WebDriver::EdgeChrome::Options
-              ::Selenium::WebDriver::EdgeChrome::Options.new
-            else
-              ::Selenium::WebDriver::Edge::Options.new
             end
           end